+++ scripts/tests/utils.py
@@ -1,4 +1,4 @@
 from ec.utils import get
 from ec.types.basics import yn
 
-print get('Say someting', yn())
+print get('Say something', yn())